modelling/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "da-models"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "pyarrow>=15",
  "holidays>=0.50",
  "python-dotenv>=1.0",
  "azure-storage-blob>=12.19",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["da_models*"]